선택한 직군에 대해 해당하는 학과 없을시 "해당 하는 학과가 없습니다" 출력
